SEQUOIADBMAINSTREAM-2650 优化： 远程 file 对象优化, 在omagent 保存已打开文件描述符 SEQUOIADBMAINSTREAM-2651 fix bug: 创建远程文件时，权限指定只读，打开模式指定读写，读写文件报权限错误
git-svn-id: http://192.168.20.11/sequoiadb/trunk@30565 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/include/sptUsrFileCommon.hpp
@@ -59,7 +59,7 @@ namespace engine
 
       INT32 write( const CHAR* buf, SINT64 size, std::string &err ) ;
 
-      INT32 seek( INT32 seekSize, const bson::BSONObj &optionObj,
+      INT32 seek( INT64 seekSize, const bson::BSONObj &optionObj,
                   std::string &err ) ;
 
       INT32 close( std::string &err ) ;
